+++ README.md
@@ -18,7 +18,7 @@ Key features include:
 ## 📁 File Tree
 
 ```
-launcher.py
+launcher.py  (CustomTkinter)
 LICENSE
 main.py
 README.md
