+++ data/config.json
@@ -233,4 +233,4 @@
       "end": "23:45"
     }
   }
-}
+
