+++ scripts/build_student_periods.py
@@ -84,3 +84,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+
