+++ src/database.py
@@ -63,3 +63,4 @@ def create_app() -> Flask:
             db.create_all()
 
     return app
+
