+++ src/routes/auth.py
@@ -158,3 +158,4 @@ def logout():
 
     session.clear()
     return redirect(url_for('auth.login'))
+
