+++ src/routes/core.py
@@ -259,3 +259,4 @@ def debug_session():
 @ping_bp.route('/ping')
 def ping():
     return "pong", 200
+
