+++ src/utils.py
@@ -132,3 +132,4 @@ def csv_response(string_io, filename="export"):
     output.headers["Content-Disposition"] = f"attachment; filename={filename}.csv"
     output.headers["Content-type"] = "text/csv"
     return output
+
