+++ wsgi.py
@@ -15,3 +15,4 @@ def get_app():
 if __name__ == "__main__":
     app = get_app()
     app.run()
+
